core/notifications/requirements.txt
requests==2.31.0
Flask==2.3.2
python-dotenv==1.0.0
setuptools==58.1.0
wheel==0.38.4 
